Addressed comment regarding /etc/waagent.conf
