+++ Utils/test/test_ovf_utils.py
@@ -7,7 +7,18 @@ import Utils.ovfutils as ovf_utils
 import Utils.logger as logger
 
 
-Configuration = ext_utils.ConfigurationProvider(None)
+# dummy configuration class based on vmaccess.Configuration
+class Configuration:
+    def __init__(self):
+        self.dictionary = {
+            "Provisioning.DecodeCustomData": "n"
+        }
+
+    def get(self, key):
+        return self.dictionary.get(key)
+
+
+config = Configuration()
 
 logger.global_shared_context_logger = logger.TestLogger()
 
@@ -16,7 +27,7 @@ class TestTestOvfUtils(unittest.TestCase):
     def test_ovf_env_parse(self):
         current_dir = d = path.dirname(path.abspath(__file__))
         ovf_xml = ext_utils.get_file_contents(path.join(current_dir, 'ovf-env.xml'))
-        ovf_env = ovf_utils.OvfEnv.parse(ovf_xml, Configuration)
+        ovf_env = ovf_utils.OvfEnv.parse(ovf_xml, config)
         self.assertIsNotNone(ovf_env, "ovf_env should not be null")
